Make get_sources_metadata_images() a static method of MetadataItem
And get rid of imports
